+++ src/main/java/de/jcup/eclipse/commons/document/MetaCharacterDetector.java
@@ -1,0 +1,7 @@
+package de.jcup.eclipse.commons.document;
+
+public interface MetaCharacterDetector {
+
+    boolean isMetaCharacter(char charBefore);
+
+}