+++ src/main/java/de/jcup/eclipse/commons/ui/DefaultColorManager.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2017 Albert Tregnaghi
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions
+ * and limitations under the License.
+ *
+ */
+package de.jcup.eclipse.commons.ui;
+
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+
+import org.eclipse.swt.graphics.Color;
+import org.eclipse.swt.graphics.RGB;
+import org.eclipse.swt.widgets.Display;
+
+public class DefaultColorManager implements ColorManager {
+
+    private Map<RGB, Color> fColorTable = new HashMap<>(10);
+
+    @Override
+    public void dispose() {
+        Iterator<Color> e = fColorTable.values().iterator();
+        while (e.hasNext()) {
+            e.next().dispose();
+        }
+    }
+
+    @Override
+    public Color getColor(RGB rgb) {
+        if (rgb == null) {
+            return null;
+        }
+        Color color = fColorTable.get(rgb);
+        if (color == null) {
+            color = new Color(Display.getCurrent(), rgb);
+            fColorTable.put(rgb, color);
+        }
+        return color;
+    }
+
+}